Encapsulate comment features in one class
