+++ portfolio/src/main/java/com/google/sps/servlets/Comment.java
@@ -16,19 +16,16 @@ package com.google.sps.servlets;
 
 /** This class sets the text and date data for a {@code Comment} */
 public class Comment {
-  private final long day;
-  private final long month;
-  private final long year;
-  private final String text;
+  private final String postedTime;
   private final String title;
+  private final String text;
   private final String imageUrl;
 
-  public Comment(String title, String text, String imageUrl, long day, long month, long year) {
+  public Comment(String postedTime, String title, String text, String imageUrl) {
+    this.postedTime = postedTime;
     this.title = title;
     this.text = text;
     this.imageUrl = imageUrl;
-    this.day = day;
-    this.month = month;
-    this.year = year;
+    
   }
 }
