+++ portfolio/src/main/java/com/google/sps/servlets/DataServlet.java
@@ -34,10 +34,14 @@ import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.gson.Gson;
 import java.io.IOException;
+import java.time.format.DateTimeFormatter;
+import java.time.format.FormatStyle;
+import java.time.LocalDateTime;
+import java.time.ZonedDateTime;
+import java.time.ZoneId;
 import java.net.MalformedURLException;
 import java.net.URL;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -54,38 +58,18 @@ import javax.servlet.http.HttpServletResponse;
 public class DataServlet extends HttpServlet {
 
   private final DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
+  private final CommentService commentService = new CommentService();
 
   @Override
   public void doGet(HttpServletRequest request, HttpServletResponse response) throws IOException {
     
-    Query query = new Query("Comment").addSort("year", SortDirection.DESCENDING)
-                                      .addSort("month", SortDirection.DESCENDING)
-                                      .addSort("day", SortDirection.DESCENDING);
-
+    Query query = new Query("Comment").addSort("postedTime", SortDirection.DESCENDING);
     PreparedQuery results = datastore.prepare(query);
-
     String amountParam = request.getParameter("comment-amount");
 
-    /** Defailt value for commentAmount is 10 in case there's a parsing error. */
+    /** Default value for commentAmount is 10 in case there's a parsing error. */
     int commentAmount = amountParam == null ? 10 : Integer.parseInt(amountParam);
-
-    Iterable<Entity> newComments = results.asIterable(FetchOptions.Builder.withLimit(commentAmount));    
-    
-    /** /data page is update to contain JSON file of all imgaes up to commentAmount. */
-    List<Comment> comments = results.asList(FetchOptions.Builder.withLimit(commentAmount))
-        .stream()
-        .map(entity -> {
-            String title = (String) entity.getProperty("title");
-            String text = (String) entity.getProperty("text");
-            String imageUrl = (String) entity.getProperty("imageUrl");
-            long day = (long) entity.getProperty("day");
-            long month = (long) entity.getProperty("month");
-            long year = (long) entity.getProperty("year");
-
-            return new Comment(title, text, imageUrl, day, month, year);
-        })
-        .collect(Collectors.toList());
-      
+    List<Comment> comments = commentService.getComments(results, commentAmount); 
     Gson gson = new Gson();
 
     response.setContentType("application/json;");
@@ -95,16 +79,7 @@ public class DataServlet extends HttpServlet {
   @Override
   public void doPost(HttpServletRequest request, HttpServletResponse response) throws IOException { 
 
-    Entity commentEntity = new Entity("Comment");
-
-    commentEntity.setProperty("imageUrl", getUploadedFileUrl(request, "image"));    
-    commentEntity.setProperty("day", Calendar.getInstance().get(Calendar.DATE));
-    commentEntity.setProperty("month", Calendar.getInstance().get(Calendar.MONTH));
-    commentEntity.setProperty("text", request.getParameter("text"));
-    commentEntity.setProperty("title", request.getParameter("title"));
-    commentEntity.setProperty("year", Calendar.getInstance().get(Calendar.YEAR));
-
-    datastore.put(commentEntity);
+    commentService.saveComment(request);
 
     response.sendRedirect("/index.html");
   }
@@ -112,58 +87,9 @@ public class DataServlet extends HttpServlet {
   @Override
   public void doDelete(HttpServletRequest request, HttpServletResponse response) throws IOException { 
 
-    DatastoreService datastore = DatastoreServiceFactory.getDatastoreService();
-
-    removeEntities("Comment");
-    removeEntities("__BlobInfo__");
-    removeEntities("__BlobUploadSession__");
+    commentService.deleteAllComments();
 
     response.sendRedirect("/index.html");
   }
 
-  /** Removes entities from datastore based off a query string. */
-  private void removeEntities(String query) {
-    datastore.prepare(new Query(query))
-             .asList(FetchOptions.Builder.withLimit(Integer.MAX_VALUE))
-             .stream()
-             .forEach(entity -> datastore.delete(entity.getKey()));                                          
-  }
-  
- /** Returns a URL that points to the uploaded file, or null if the user didn't upload a file. */
-  private String getUploadedFileUrl(HttpServletRequest request, String formInputElementName) {
-    BlobstoreService blobstoreService = BlobstoreServiceFactory.getBlobstoreService();
-    Map<String, List<BlobKey>> blobs = blobstoreService.getUploads(request);
-    List<BlobKey> blobKeys = blobs.get("image");
-
-    // User submitted form without selecting a file, so we can't get a URL. (dev server)
-    if (blobKeys == null || blobKeys.isEmpty()) {
-      return null;
-    }
-
-    // Our form only contains a single file input, so get the first index.
-    BlobKey blobKey = blobKeys.get(0);
-
-    // User submitted form without selecting a file, so we can't get a URL. (live server)
-    BlobInfo blobInfo = new BlobInfoFactory().loadBlobInfo(blobKey);
-    if (blobInfo.getSize() == 0) {
-      blobstoreService.delete(blobKey);
-      return null;
-    }
-
-    // We could check the validity of the file here, e.g. to make sure it's an image file
-    // https://stackoverflow.com/q/10779564/873165
-
-    // Use ImagesService to get a URL that points to the uploaded file.
-    ImagesService imagesService = ImagesServiceFactory.getImagesService();
-    ServingUrlOptions options = ServingUrlOptions.Builder.withBlobKey(blobKey);
-
-    // To support running in Google Cloud Shell with AppEngine's devserver, we must use the relative
-    // path to the image, rather than the path returned by imagesService which contains a host.
-    try {
-      URL url = new URL(imagesService.getServingUrl(options));
-      return url.getPath();
-    } catch (MalformedURLException e) {
-      return imagesService.getServingUrl(options);
-    }
-  }
 }
