+++ portfolio/src/main/webapp/index.html
@@ -10,10 +10,10 @@
       professional and academic information about myself.
       Clicking on the buttons below will take you to 
       different parts of my life. </h3>
-  <button id="header-login">Login</button>
+  <a href="/auth" id="header-login">Login</a>
   <div class="hidden" id="logout-container"> 
     <p id="user-display"></p>
-    <button id="logout">Logout</button>
+    <a href="/auth">Logout</a>
   </div>
 </div>
 <div id="layout-container">
@@ -87,8 +87,8 @@
       <input id="comment-amount" min="1" max="50" name="comment-amount" type="number" value="10">
     </form>
     <button id="delete-comments">Delete Comments</button>
-    <button id="login">Log in to leave comment</button> 
-    <button class="hidden" id="to-comment-page">Post a Comment</button>
+    <a href="/auth" id="login">Log in to leave comment</a> 
+    <a class="hidden" href="/comment-constructor.jsp" id="to-comment-page">Post a comment</a>
     <hr>
     <div id="comments"></div>
   </div>
